pyproject.toml
[tool.poetry]
name = "synapp"
version = "0.1.0"
description = "Library for interfacing with EEG devices, and recording and processing data"
authors = ["Logan <logandanek@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"


[tool.poetry.dependencies]
python = "3.11.*"
numpy = "1.23.5"
scipy = "^1.11"
tensorflow = { version = "^2.15.0", optional = true }
pylsl = "^1.16.2"
muselsl = "^2.1.0"
pygatt = "4.0.5"
PyQt5 = { version = "^5.15", optional = true }
PyQt5-Qt5 = { version = "^5.15,!=5.15.11", optional = true }
matplotlib = "*"
brainflow = "*"
colorama = "^0.4.6"
pyyaml = "*"
nest-asyncio = "*"
tqdm = "*"

[tool.poetry.group.app.dependencies]
pynput = "^1.7.6"


[tool.poetry.group.dev.dependencies]
jupyter = "*"
ipykernel = "^6.29.0"
pytest = "^7.4.3"
GitPython = "*"


[tool.poetry.extras]
qt5 = ["PyQt5", "PyQt5-Qt5"]
qt6 = ["PyQt6"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

line-length = 120
indent-width = 4

# Assume Python 3.11
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
